+++ packages/generator-teams/src/app/manifestGeneration/ManifestGeneratorFactory.ts
@@ -5,52 +5,25 @@
 import { ManifestVersions } from "./ManifestVersions";
 
 import { ManifestGenerator as ManifestGenerator18 } from "./manifestGenerators/generator18/ManifestGenerator";
+import { ManifestGenerator as ManifestGenerator19 } from "./manifestGenerators/generator19/ManifestGenerator";
 import { ManifestGenerator as ManifestGeneratorDevPreview } from "./manifestGenerators/generatorDevPreview/ManifestGenerator";
 import { BaseManifestGenerator } from "./BaseManifestGenerator";
 
 export class ManifestGeneratorFactory {
 
     public static supportedManifestVersions = [
-        // {
-        //     manifestVersion: ManifestVersions.v13,
-        //     schemaUrl: "https://developer.microsoft.com/en-us/json-schemas/teams/v1.3/MicrosoftTeams.schema.json",
-        //     manifestValue: "1.3",
-        //     default: false,
-        //     hide: true
-        // },
-        // {
-        //     manifestVersion: ManifestVersions.v14,
-        //     schemaUrl: "https://developer.microsoft.com/en-us/json-schemas/teams/v1.4/MicrosoftTeams.schema.json",
-        //     manifestValue: "1.4",
-        //     default: false,
-        //     hide: true
-        // },
-        // {
-        //     manifestVersion: ManifestVersions.v15,
-        //     schemaUrl: "https://developer.microsoft.com/en-us/json-schemas/teams/v1.5/MicrosoftTeams.schema.json",
-        //     manifestValue: "1.5",
-        //     default: false,
-        //     hide: true
-        // },
-        // {
-        //     manifestVersion: ManifestVersions.v16,
-        //     schemaUrl: "https://developer.microsoft.com/en-us/json-schemas/teams/v1.6/MicrosoftTeams.schema.json",
-        //     manifestValue: "1.6",
-        //     default: false,
-        //     hide: true
-        // },
-        // {
-        //     manifestVersion: ManifestVersions.v17,
-        //     schemaUrl: "https://developer.microsoft.com/en-us/json-schemas/teams/v1.7/MicrosoftTeams.schema.json",
-        //     manifestValue: "1.7",
-        //     default: false,
-        //     hide: true
-        // },
         {
             manifestVersion: ManifestVersions.v18,
             schemaUrl: "https://developer.microsoft.com/en-us/json-schemas/teams/v1.8/MicrosoftTeams.schema.json",
             manifestValue: "1.8",
             default: true,
+            hide: false
+        },
+        {
+            manifestVersion: ManifestVersions.v19,
+            schemaUrl: "https://developer.microsoft.com/en-us/json-schemas/teams/v1.9/MicrosoftTeams.schema.json",
+            manifestValue: "1.9",
+            default: false,
             hide: false
         },
         {
@@ -79,24 +52,26 @@ export class ManifestGeneratorFactory {
 
     public createManifestGenerator(manifestVersion: string): BaseManifestGenerator {
         if (manifestVersion == ManifestVersions.v18) {
-            return new ManifestGenerator18();            
+            return new ManifestGenerator18();
+        } else if (manifestVersion == ManifestVersions.v19) {
+            return new ManifestGenerator19();
         } else if (manifestVersion == ManifestVersions.devPreview) {
             return new ManifestGeneratorDevPreview();
-        }else {
+        } else {
             throw new Error(`Invalid manifest version: ${manifestVersion}.`);
         }
     }
 
     static isSchemaVersionValid(existingManifest: any): boolean {
         if (existingManifest) {
-            let retval = false;
+            let retVal = false;
             this.supportedManifestVersions.forEach(manifestVersionItem => {
                 if (existingManifest["$schema"] == manifestVersionItem.schemaUrl) {
-                    retval = true;
+                    retVal = true;
                     return;
                 }
             });
-            return retval;
+            return retVal;
         } else {
             return true;
         }