+++ packages/generator-teams/src/app/manifestGeneration/ManifestVersions.ts
@@ -3,11 +3,7 @@
 // Licensed under the MIT license.
 
 export enum ManifestVersions {
-    // v13 = "v1.3",
-    // v14 = "v1.4",
-    // v15 = "v1.5",
-    // v16 = "v1.6",
-    // v17 = "v1.7",
     v18 = "v1.8",
+    v19 = "v1.9",
     devPreview = "devPreview"
 }